fix(atlas): accumulate sprite area in 64-bit
m_area summed (width * height) in uint32, which overflows once the
combined sprite area exceeds ~4.29e9 px (reachable with large batches in
multi-atlas mode). The truncated area produced a too-small start size in
calcSize() and a wrong reported fill percentage. Widen m_area and
getArea() to uint64 and cast the multiply; thread the wider type through
saveAtlas().

+++ src/Atlas/AtlasSize.cpp
@@ -35,10 +35,10 @@ void cAtlasSize::addRect(const sSize& size)
     m_maxRectSize.width = std::max(m_maxRectSize.width, width);
     m_maxRectSize.height = std::max(m_maxRectSize.height, height);
 
-    m_area += width * height;
+    m_area += static_cast<uint64_t>(width) * height;
 }
 
-uint32_t cAtlasSize::getArea() const
+uint64_t cAtlasSize::getArea() const
 {
     return m_area;
 }
@@ -50,7 +50,7 @@ sSize cAtlasSize::calcSize() const
     auto h = std::max(sq, m_maxRectSize.height) + m_config.border * 2u;
 
     auto width = std::max(w, h);
-    auto height = m_area / width;
+    auto height = static_cast<uint32_t>(m_area / width);
 
     return {
         FixSize(width, m_config.pot),
