+++ src/ImageList.h
@@ -54,7 +54,7 @@ private:
 
     bool saveAtlas(AtlasPacker* packer, const char* atlasName,
                    const char* resPathPrefix, cFile& xmlFile,
-                   const sSize& atlasSize, uint32_t spritesArea, uint64_t startTime);
+                   const sSize& atlasSize, uint64_t spritesArea, uint64_t startTime);
 
     bool findBestSortAndSize(ImageList& images, const sSize& startSize, const sSize& maxSize, sSize& outSize);
     bool findMinimalAtlasSize(AtlasPacker* packer, ImageList& images, const sSize& startSize, sSize& outSize);
